feat: migrate aws-kms-kotlin-base WorkActions to DefaultTasks
